test/requirements.txt
iso8601>=0.1.10
http_parser>=0.8.3
msgpack-python>=0.4.6
pyjwt>=1.1.0
requests>=2.4.0
six>=1.9.0
wsaccel>=0.6.2
